Refactoring the support messeges
