+++ js/script.js
@@ -23,6 +23,10 @@ import { readSingleFile } from "./scripts/dataExchangeFunctions.js";
 // Importing Ship3D library
 import { Ship3D } from "../libs/3D_engine/Ship3D.js";
 
+// Supporting functions
+import { showMessage } from "./scripts/supportFunctions.js";
+import { assignColorToComponent } from "./scripts/supportFunctions.js";
+
 // Basic Three.js setup
 let zUpCont, scene, camera, renderer;
 
@@ -174,18 +178,6 @@ function onMouseMove( event ) {
 
 }
 
-// Function to check if a color is dark
-function isColorDark(color) {
-    // input a color THREE.Color element
-
-    // Use the luminance formula to determine brightness
-    let luminance = 0.299 * color.r + 0.587 * color.g + 0.114 * color.b;
-
-    // Return true if the color is dark
-    return luminance < 0.128;
-
-}
-
 function onMouseDoubleClick (event) {
 
     const selectedName = document.getElementById('selectedName')
@@ -235,29 +227,6 @@ function onMouseDoubleClick (event) {
 
 }
 
-// Generic function for error shown display
-function showMessage(errorMessageText) {
-    const errorContainer = document.getElementById('errorContainer');
-    const errorMessage = document.createElement('div');
-
-    errorMessage.textContent = errorMessageText;
-    errorMessage.className = 'bg-red-600 text-white px-4 py-2 rounded opacity-0 transition-opacity duration-3000';
-
-    // Append the new error message to the container
-    errorContainer.appendChild(errorMessage);
-
-    // Show the error message
-    setTimeout(() => {
-        errorMessage.classList.replace('opacity-0', 'opacity-100');
-    }, 10); // Delay to ensure the element is rendered before starting the animation
-
-    // Hide and remove the error message after 3 seconds
-    setTimeout(() => {
-        errorMessage.classList.replace('opacity-100', 'opacity-0');
-        setTimeout(() => errorMessage.remove(), 1000); // Remove after fade-out
-    }, 3000);
-}
-
 // Event listeners for creating and deleting blocks
 document.getElementById('create-block').addEventListener('click', () => {
     
@@ -305,7 +274,7 @@ document.getElementById('delete-block').addEventListener('click', () => {
     
     
     if ( !elementClicked ) {
-        showMessage("Error: No element selected");
+        showMessage();
         return
     }    
     
@@ -353,24 +322,10 @@ document.getElementById('delete-block').addEventListener('click', () => {
 
 });
 
-function assignColorToComponent(component, valueString) {
-    /* 
-    Function receive a component and an color string
-    if the color is dark the text will appear as white and placeholder as gray
-    */
-
-    const color = new THREE.Color(valueString)
-    component.style.backgroundColor = "#" +  color.getHexString();
-
-    isColorDark(color) ? component.classList.add('placeholder-gray-100') : component.classList.remove('placeholder-gray-100')
-    isColorDark(color) ? component.classList.add('text-white') : component.classList.remove('text-white')
-
-}
-
 function changeTankColor (valueString, elementClickedName) {
 
     if(!elementClicked){
-        showMessage("Error: No Element Selected");
+        showMessage();
         return
     }
 
@@ -399,7 +354,7 @@ function changeTankColor (valueString, elementClickedName) {
 function changeVariableValue(valueString,  dimension, elementClickedName) {
     
     if(!elementClicked){
-        showMessage("Error: No Element Selected");
+        showMessage();
         return
     }
